refactor props and file names
